Added runtime-fs.js
